% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof_path.R
\name{assert_pprof}
\alias{assert_pprof}
\title{Check if \code{pprof} is installed.}
\usage{
assert_pprof(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, whether to print out messages
when \code{proffer} is having trouble searching for \code{pprof}.}
}
\description{
Returns silently if \code{pprof} is installed
and throws an error if \code{pprof} is missing.
}
\details{
See \url{https://github.com/r-prof/proffer#installation}
for setup instructions.
}
\examples{
\dontrun{
assert_pprof()
}
}
