% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_go}
\alias{install_go}
\title{Deprecated}
\usage{
install_go(destination = tempfile(), version = NULL, quiet = FALSE)
}
\arguments{
\item{destination}{Full path to the Go installation with the \code{pprof}
and Go executables.}

\item{version}{Character, a version string such as \code{"1.19.5"}.}

\item{quiet}{Logical, whether to suppress console messages.}
}
\description{
Deprecated. To install Go, visit \url{https://go.dev/dl/}
and choose the appropriate build for your platform.
}
\keyword{internal}
