% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{pprof_path}
\alias{pprof_path}
\title{Deprecated}
\usage{
pprof_path()
}
\value{
Character, path to Go.
}
\description{
Deprecated.
}
\examples{
if (identical(Sys.getenv("PROFFER_EXAMPLES"), "true")) {
pprof_path()
}
}
\keyword{internal}
