% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_pprof.R
\name{serve_pprof}
\alias{serve_pprof}
\title{Visualize profiling data with pprof.}
\usage{
serve_pprof(
  pprof,
  host = "localhost",
  port = NULL,
  browse = interactive(),
  verbose = TRUE
)
}
\arguments{
\item{pprof}{Path to pprof samples.}

\item{host}{Host name. Set to \code{"localhost"} to view locally
or \code{"0.0.0.0"} to view from another machine.}

\item{port}{Port number.}

\item{browse}{Logical, whether to open a browser to view
the pprof server.}

\item{verbose}{Logical, whether to print the URL of the pprof
server to the R console as a message.}
}
\value{
A \code{callr::r_bg()} handler. Use this handler
}
\description{
Visualize profiling data with pprof.
}
\details{
Uses a local interactive server.
Navigate a browser to a URL in the message.
The server starts in a background process
}
\examples{
\dontrun{
pprof <- record_pprof(replicate(1e2, sample.int(1e4)))
# Start a pprof virtual server in the background.
px <- serve_pprof(pprof)
# Terminate the server.
px$kill()
}
}
