% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof.R
\name{serve_rprof}
\alias{serve_rprof}
\title{Visualize Rprof() output with pprof.}
\usage{
serve_rprof(
  rprof,
  host = "localhost",
  port = NULL,
  browse = interactive(),
  verbose = TRUE
)
}
\value{
A \code{callr::r_bg()} handle. Use this handle
to take down the server with \verb{$kill()}.
}
\description{
Use pprof to visualize profiling data
produced by \code{Rprof()} or \code{\link[=record_rprof]{record_rprof()}}.
}
\details{
Uses a local interactive server.
Navigate a browser to a URL in the message.
The server starts in a background process
}
\examples{
\dontrun{
rprof <- record_rprof(replicate(1e2, sample.int(1e4)))
# Start a pprof virtual server in the background.
px <- serve_rprof(rprof)
# Terminate the server.
px$kill()
}
}
