% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/troubleshoot.R
\name{test_pprof}
\alias{test_pprof}
\title{Test \code{pprof()}}
\usage{
test_pprof(
  host = "localhost",
  port = NULL,
  browse = interactive(),
  verbose = TRUE
)
}
\arguments{
\item{host}{Host name. Set to \code{"localhost"} to view locally
or \code{"0.0.0.0"} to view from another machine. If you view
from another machine, the printed out URL will not be valid.
For example, if \code{pprof()} or \code{serve_pprof()} prints
"http://0.0.0.0:8080", then you need to replace 0.0.0.0
with your computer's name or IP address, e.g.
"http://my_computer.com:8080".}

\item{port}{Port number for hosting the local pprof server.
Chosen randomly by default.}

\item{browse}{Logical, whether to open a browser to view
the pprof server.}

\item{verbose}{Logical, whether to print console messages
such as the URL of the local \code{pprof} server.}
}
\description{
Do a test run of \code{pprof()} to verify that the
system dependencies like \code{pprof} work as expected.
}
\details{
See \url{https://github.com/r-prof/proffer#installation}
for setup instructions.
}
\examples{
\dontrun{
test_pprof()
}
}
\seealso{
\code{\link[=pprof]{pprof()}}
}
