% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_pprof.R
\name{to_pprof}
\alias{to_pprof}
\title{Convert Rprof samples to pprof format.}
\usage{
to_pprof(rprof, pprof = tempfile())
}
\arguments{
\item{rprof}{Path to Rprof samples.}

\item{pprof}{Path to pprof samples.}
}
\value{
Path to pprof samples.
}
\description{
Convert Rprof samples to pprof format.
}
\examples{
rprof <- record_rprof(replicate(1e2, sample.int(1e4)))
to_pprof(rprof)
}
