% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record.R
\name{to_rprof}
\alias{to_rprof}
\title{Convert pprof samples to Rprof format.}
\usage{
to_rprof(pprof, rprof = tempfile())
}
\arguments{
\item{pprof}{Path to pprof samples.}

\item{rprof}{Path to Rprof samples.}
}
\value{
Path to pprof samples.
}
\description{
Convert pprof samples to Rprof format.
}
\examples{
pprof <- record_pprof(replicate(1e2, sample.int(1e4)))
to_rprof(pprof)
}
