% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{vis_pprof}
\alias{vis_pprof}
\title{Visualize profiling data with pprof.}
\usage{
vis_pprof(pprof, host = "localhost", port = NULL)
}
\arguments{
\item{pprof}{Path to pprof samples.}

\item{host}{Host name. Set to \code{"localhost"} to view locally
or \code{"0.0.0.0"} to view from another machine.}

\item{port}{Port number.}
}
\value{
Nothing.
}
\description{
Visualize profiling data with pprof.
}
\details{
Uses a local interactive server.
Navigate a browser to a URL in the message.
}
\examples{
pprof <- record_pprof(replicate(1e2, sample.int(1e4)))
to_rprof(pprof)
}
