% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_go.R
\name{install_go}
\alias{install_go}
\title{Install pprof and Go}
\usage{
install_go(destination = Sys.getenv("HOME"), version = "1.14", quiet = FALSE)
}
\arguments{
\item{destination}{Full path to install pprof and Go on Linux systems.
Defaults to \code{Sys.getenv("HOME")}, which means the default
Go installation path is \code{file.path(Sys.getenv("HOME"), "go")}.
If you are using Linux, after the installation succeeds,
follow the instructions in the
console messages for setting the appropriate environment
variables so \code{proffer} can find \code{pprof} and Go.}

\item{version}{Character, a version string such as \code{"1.14"}}
}
\description{
This function installs Go, which comes with its own
installation of \code{pprof}. Only for amd64 architectures running
Linux, Mac OS, or Windows.
}
\details{
On Linux, users will need to set
the environment variables \code{PROFFER_PPROF_BIN} and \code{PROFFER_GO_BIN} using
\code{usethis::edit_r_environ()}. Typically, if \code{destination} is \verb{/home/you},
then typically those lines look like
\verb{PROFFER_GO_BIN=/home/you/go/pkg/tool/linux_amd64/pprof}
\verb{PROFFER_PPROF_BIN=/home/you/go/bin/go}.
}
