% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{pprof_path}
\alias{pprof_path}
\title{Show the path to the pprof executable.}
\usage{
pprof_path()
}
\value{
Character, path to \code{pprof} it exists and \code{""} otherwise.
}
\description{
Defaults to the \code{PROFFER_PPROF_BIN} environment variable.
Otherwise, it searches your Go lang installation for \code{pprof}.
}
\details{
See \url{https://github.com/r-prof/proffer#installation}
for setup instructions.
}
\examples{
if (identical(Sys.getenv("PROFFER_EXAMPLES"), "true")) {
pprof_path()
}
}
