% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof_path.R
\name{pprof_path}
\alias{pprof_path}
\title{Show the path to the pprof executable.}
\usage{
pprof_path(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, whether to print out messages
when \code{proffer} is having trouble searching for \code{pprof}.}
}
\value{
Character, path to \code{pprof} it exists and \code{""} otherwise.
}
\description{
Defaults to the \code{pprof_path} environment variable.
Otherwise, it searches your Go lang installation for \code{pprof}.
}
\details{
See \url{https://github.com/r-prof/proffer#installation}
for setup instructions.
}
\examples{
pprof_path()
}
