% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{proffer-package}
\alias{proffer-package}
\alias{proffer}
\title{proffer: pprof-powered profiling packaged for R}
\description{
Profiling in R powered by pprof.
}
\examples{
# TBD
\dontrun{
# Start a pprof virtual server in the background.
px <- pprof(replicate(1e2, sample.int(1e4)))
# Terminate the server.
px$kill()
}
}
\references{
\url{https://github.com/wlandau/pprof}
}
\author{
William Michael Landau \email{will.landau@gmail.com}
}
