% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprof.R
\name{random_port}
\alias{random_port}
\title{Choose a random TCP port.}
\usage{
random_port(lower = 49152L, upper = 65355L)
}
\arguments{
\item{lower}{Integer of length 1, lower bound of the port number.}

\item{upper}{Integer of length 1, upper bound of the port number.}
}
\value{
Port number, positive integer of length 1.
}
\description{
Choose a random TCP port that is unlikely to be occupied
by another process.
}
