% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{record_pprof}
\alias{record_pprof}
\title{Profile R code and record pprof samples.}
\usage{
record_pprof(expr, pprof = tempfile())
}
\arguments{
\item{expr}{An R expression to profile.}

\item{pprof}{Path to a file with pprof samples.
Also returned from the function.}
}
\value{
Path to a file with pprof samples.
}
\description{
Run R code and record pprof samples.
}
\examples{
# Returns a path to pprof samples.
record_pprof(replicate(1e2, sample.int(1e4)))
}
