% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.R
\name{record_rprof}
\alias{record_rprof}
\title{Profile R code and record Rprof samples.}
\usage{
record_rprof(expr)
}
\arguments{
\item{expr}{An R expression to profile.}

\item{pprof}{Path to a file with Rprof samples.
Also returned from the function.}
}
\value{
Path to a file with Rprof samples.
}
\description{
Run R code and record Rprof samples.
}
\examples{
# Returns a path to Rprof samples.
record_rprof(replicate(1e2, sample.int(1e4)))
}
